% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_functions.R
\name{myplot}
\alias{myplot}
\title{Plot Output}
\usage{
myplot(output)
}
\arguments{
\item{output}{some output}
}
\description{
Some plotting function that displays output from an OM/EM run using lasac
}
\details{
Further details go here
}
