% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_functions.R
\name{rich_sell}
\alias{rich_sell}
\title{Rich's selectivity-at-length equation}
\usage{
rich_sell(ages, lbins, sel_type, sel_params, growth, natM)
}
\arguments{
\item{ages}{vector of ages required to translate selectivity by age to selectivity by length}

\item{lbins}{the length bins required}

\item{sel_type}{character of type "double_normal" etc. consistent with earthfish syntax}

\item{sel_params}{list in earthfish syntax appropriate for sel_type}

\item{growth}{list of VB growth parameters}

\item{natM}{natural mortality}
}
\description{
A function for determining selectivity at length bins given selectivity at age and von Bertalanffy growth parameters. Also uses natural mortality.
}
